Переместил часть самых базовых классов в новую библиотеку QS.Project.Core
Собранную под NetStandart2.0
