+++ Modules/QS.BusinessCommon/QS.BusinessCommon.csproj
@@ -47,6 +47,10 @@
     <Folder Include="Repository\" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\QS.Project.Core\QS.Project.Core.csproj">
+      <Project>{c01ae5a4-3cff-4c7b-884d-45b07acde761}</Project>
+      <Name>QS.Project.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\QS.Project\QS.Project.csproj">
       <Project>{28A78143-42E3-40E7-86F5-DD7E1679BE86}</Project>
       <Name>QS.Project</Name>
