+++ QS.LibsTest/QS.LibsTest.csproj
@@ -181,6 +181,10 @@
       <Project>{D21C0F94-5B1B-49E1-A5B0-50C3F99287E2}</Project>
       <Name>QSBanks</Name>
     </ProjectReference>
+    <ProjectReference Include="..\QS.Project.Core\QS.Project.Core.csproj">
+      <Project>{c01ae5a4-3cff-4c7b-884d-45b07acde761}</Project>
+      <Name>QS.Project.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\QS.Project\QS.Project.csproj">
       <Project>{28A78143-42E3-40E7-86F5-DD7E1679BE86}</Project>
       <Name>QS.Project</Name>
