+++ QS.Project.Core/Dialog/WindowGravity.cs
