+++ QS.Project.Core/Navigation/CloseSource.cs
