+++ QS.Project.Core/QS.Project.Core.csproj
@@ -7,4 +7,9 @@
         <LangVersion>8</LangVersion>
     </PropertyGroup>
 
+    <ItemGroup>
+      <PackageReference Include="Autofac" Version="6.2.0" />
+      <PackageReference Include="NHibernate" Version="5.3.8" />
+    </ItemGroup>
+
 </Project>
