+++ QS.Project.Core/ViewModels/Dialog/DialogViewModelBase.cs
