+++ QS.Project.Core/ViewModels/Dialog/WindowDialogViewModelBase.cs
