+++ QS.Project.Core/ViewModels/Extension/IOnCloseActionViewModel.cs
