+++ QS.Project.Core/ViewModels/Extension/IWindowDialogSettings.cs
