+++ QS.Project.Core/ViewModels/ViewModelBase.cs
